Extracted TimePickerBaseSelectionChangedEventArgs<T> into a separate file
